Add files via upload
visual shows how the top 10 countries which have opportunities for growth are
performing. This data does not include the UK as the country already has high demand and I’ve
been told you’re more focused on the countries where demand can be increased. The analysis
shows that countries such as the Netherlands, Ireland, Germany and France have high volumes
of units bought and revenue generated. I would suggest that these countries should be focused
on to ensure that measures are taken to capture these markets even more.