Add files via upload
analysis has been performed on the top 10 customers who have purchased the most
from the store. The data shows that there is not much of a difference between the purchases
made by the top 10 customers. The highest revenue generating customer only purchased 17%
more than the 2nd highest which shows that the business is not relying only on a few customers
to generate the revenue. This shows that the bargaining power of customers is low and the
business is in a good position.