Global Product Demand by Country (Excluding UK)
the map chart shows the regions that have generated the most revenue compared with
the regions that have not. It can be seen that apart from the UK, countries such as Netherlands,
Ireland, Germany, France and Australia are generating high revenue and the company should
invest more in these areas to increase demand for products. The map also shows that most of
the sales are only in the European region with very few in the American region. Africa and Asia
do not have any demand for the products, along with Russia. A new strategy targeting these
areas has the potential to boost sales revenues and profitability.